+++ webpack.config.js
@@ -21,7 +21,7 @@ module.exports = () => {
 
     config.plugins.push(
         mf_config({
-            name: "plone-mosaic",
+            name: package_json.name,
             filename: "plone-mosaic-remote.min.js",
             remote_entry: config.entry["plone-mosaic.min"],
             dependencies: {